# gofile/apiserver/handler/user/delete.go
package user

import (
	"strconv"

	"../../handler"
	"../../model"
	"../../pkg/errno"
	"github.com/gin-gonic/gin"
)

//Delete is to delete user info
// @Summary Delete an user by the user identifier
// @Description Delete user by ID
// @Tags user
// @Accept  json
// @Produce  json
// @Param id path uint64 true "The user's database id index num"
// @Success 200 {object} handler.Response "{"code":0,"message":"OK","data":null}"
// @Router /v1/user/{id} [delete]
func Delete(c *gin.Context) {
	userID, _ := strconv.Atoi(c.Param("id"))
	if err := model.DeleteUser(uint64(userID)); err != nil {
		handler.SendResponse(c, errno.ErrDatabase, nil)
		return
	}
	handler.SendResponse(c, nil, nil)
}